requirements.txt
fpdf
google-genai
langchain
langchain-community
langchain-core
matplotlib
nltk
numpy
openai
pandas
plotly
python-dotenv
requests
scikit-learn
streamlit
ta
xgboost
yfinance
google-generativeai
